zero-trust-gateway/requirements.txt

Flask==2.3.3
redis==5.0.0
PyJWT==2.8.0
requests==2.31.0
prometheus-client==0.17.1


openziti==0.8.1


pandas==2.1.0
matplotlib==3.7.2


gunicorn==21.2.0
python-dotenv==1.0.0